split off Alignment tests from pairwise aligner tests
